# Optimize traffic sql hint algorithm performance