--- a/shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/hint/SQLHintPropertiesKey.java
+++ b/shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/hint/SQLHintPropertiesKey.java
@@ -36,7 +36,12 @@ public enum SQLHintPropertiesKey implements TypedPropertyKey {
     /**
      * Whether hint route write data source or not.
      */
-    WRITE_ROUTE_ONLY_KEY("writeRouteOnly", String.valueOf(Boolean.FALSE), boolean.class);
+    WRITE_ROUTE_ONLY_KEY("writeRouteOnly", String.valueOf(Boolean.FALSE), boolean.class),
+    
+    /**
+     * Whether to use traffic or not.
+     */
+    USE_TRAFFIC("useTraffic", String.valueOf(Boolean.FALSE), boolean.class);
     
     private final String key;